Refactor to share test
