+++ steps/share/context.go
@@ -1,22 +1,22 @@
-package sampletest
+package share
 
 import (
 	"github.com/cucumber/godog"
 	"github.com/owncloud/cs3api-validator/featurecontext"
 )
 
-// SampleTestFeatureContext holds values which are used across test steps
-type SampleTestFeatureContext struct {
+// ShareTestFeatureContext holds values which are used across test steps
+type ShareTestFeatureContext struct {
 	*featurecontext.FeatureContext
 }
 
-func NewSampleTestFeatureContext(fc *featurecontext.FeatureContext, sc *godog.ScenarioContext) *SampleTestFeatureContext {
-	nsc := &SampleTestFeatureContext{FeatureContext: fc}
+func NewShareTestFeatureContext(fc *featurecontext.FeatureContext, sc *godog.ScenarioContext) *ShareTestFeatureContext {
+	nsc := &ShareTestFeatureContext{FeatureContext: fc}
 	nsc.Register(sc)
 	return nsc
 }
 
-func (f *SampleTestFeatureContext) Register(ctx *godog.ScenarioContext) {
+func (f *ShareTestFeatureContext) Register(ctx *godog.ScenarioContext) {
 	// steps
 	ctx.Step(`^user "([^"]*)" shares a file "([^"]*)" with the following users concurrently$`, f.UserSharesAFileWithTheFollowingUsers)
 	ctx.Step(`^the concurrent user sharing should have been successfull$`, f.TheConcurrentUserSharingShouldHaveBeenSuccessfull)
